refactor: added gesture handler internal context
