+++ example/src/App.tsx
@@ -16,6 +16,7 @@ const App = () => {
   return (
     <ShowcaseApp
       name="Bottom Sheet"
+      initialScreen="Advanced/CustomHandling"
       description={description}
       version={version}
       author={author}
