+++ src/components/bottomSheetGestureHandlersProvider/types.d.ts
@@ -1,0 +1,7 @@
+import type { ReactChild } from 'react';
+import type { BottomSheetProps } from '../bottomSheet/types';
+
+export interface BottomSheetGestureHandlersProviderProps
+  extends Pick<BottomSheetProps, 'gestureEventsHandlersHook'> {
+  children: ReactChild | ReactChild[];
+}
