+++ src/components/bottomSheetHandleContainer/types.d.ts
@@ -2,7 +2,7 @@ import type { PanGestureHandlerProperties } from 'react-native-gesture-handler';
 import type Animated from 'react-native-reanimated';
 import type { BottomSheetProps } from '../bottomSheet';
 import type { BottomSheetHandleProps } from '../bottomSheetHandle';
-import type { useInteractivePanGestureHandlerConfigs } from '../../hooks/useInteractivePanGestureHandler';
+import type { useInteractivePanGestureHandlerConfigs } from '../../hooks/useGestureHandler';
 
 export interface BottomSheetHandleContainerProps
   extends Pick<PanGestureHandlerProperties, 'simultaneousHandlers'>,
@@ -15,6 +15,5 @@ export interface BottomSheetHandleContainerProps
       | 'keyboardBehavior'
     >,
     BottomSheetHandleProps {
-  handlePanGestureHandler: any;
   handleHeight: Animated.SharedValue<number>;
 }
