+++ src/hooks/index.ts
@@ -1,17 +1,24 @@
 export { useBottomSheet } from './useBottomSheet';
 export { useBottomSheetInternal } from './useBottomSheetInternal';
-export { useScrollable } from './useScrollable';
-export { useScrollEventHandlerDefault } from './useScrollEventHandlerDefault';
 
 // modal
 export { useBottomSheetModal } from './useBottomSheetModal';
 export { useBottomSheetModalInternal } from './useBottomSheetModalInternal';
 
+// scrollable
+export { useScrollable } from './useScrollable';
+export { useScrollableSetter } from './useScrollableSetter';
+export { useScrollHandler } from './useScrollHandler';
+
+// gestures
+export { useGestureHandler } from './useGestureHandler';
+export { useGestureEventsHandlersDefault } from './useGestureEventsHandlersDefault';
+
 // utilities
 export { useKeyboard } from './useKeyboard';
 export { useStableCallback } from './useStableCallback';
-export { useInteractivePanGestureHandler } from './useInteractivePanGestureHandler';
 export { usePropsValidator } from './usePropsValidator';
 export { useNormalizedSnapPoints } from './useNormalizedSnapPoints';
 export { useReactiveSharedValue } from './useReactiveSharedValue';
 export { useBottomSheetDynamicSnapPoints } from './useBottomSheetDynamicSnapPoints';
+export { useBottomSheetGestureHandlers } from './useBottomSheetGestureHandlers';
