--- src/hooks/useScrollEventHandlerDefault.ts
@@ -1,149 +1,0 @@
-import { RefObject } from 'react';
-import { NativeScrollEvent, NativeSyntheticEvent } from 'react-native';
-import {
-  useAnimatedRef,
-  useAnimatedScrollHandler,
-  useSharedValue,
-  scrollTo,
-  useAnimatedProps,
-} from 'react-native-reanimated';
-import { useBottomSheetInternal } from './useBottomSheetInternal';
-import type { Scrollable } from '../types';
-import {
-  ANIMATION_STATE,
-  SCROLLABLE_DECELERATION_RATE_MAPPER,
-  SCROLLABLE_STATE,
-  SCROLLABLE_TYPE,
-  SHEET_STATE,
-} from '../constants';
-import { useHandleSettingScrollable } from './useHandleSettingScrollable';
-
-type HandleScrollEventContextType = {
-  initialContentOffsetY: number;
-  shouldLockInitialPosition: boolean;
-};
-
-export type UseScrollableType = (
-  type: SCROLLABLE_TYPE,
-  refreshable: boolean
-) => {
-  scrollableRef: RefObject<Scrollable>;
-  scrollableAnimatedProps: Record<string, any>;
-  handleScrollEvent: (event: NativeSyntheticEvent<NativeScrollEvent>) => void;
-  handleSettingScrollable: () => void;
-};
-
-export const useScrollEventHandlerDefault: UseScrollableType = (
-  type,
-  refreshable
-) => {
-  // refs
-  const scrollableRef = useAnimatedRef<Scrollable>();
-  const scrollableContentOffsetY = useSharedValue<number>(0);
-
-  // hooks
-  const {
-    animatedSheetState,
-    animatedScrollableState,
-    animatedAnimationState,
-    scrollableContentOffsetY: rootScrollableContentOffsetY,
-  } = useBottomSheetInternal();
-
-  // variables
-  const scrollableAnimatedProps = useAnimatedProps(() => ({
-    decelerationRate:
-      SCROLLABLE_DECELERATION_RATE_MAPPER[animatedScrollableState.value],
-    showsVerticalScrollIndicator:
-      animatedScrollableState.value === SCROLLABLE_STATE.UNLOCKED,
-  }));
-
-  // callbacks
-  const handleScrollEvent =
-    useAnimatedScrollHandler<HandleScrollEventContextType>({
-      onBeginDrag: ({ contentOffset: { y } }: NativeScrollEvent, context) => {
-        scrollableContentOffsetY.value = y;
-        rootScrollableContentOffsetY.value = y;
-        context.initialContentOffsetY = y;
-
-        /**
-         * if sheet position not extended or fill parent and the scrollable position
-         * not at the top, then we should lock the initial scrollable position.
-         */
-        if (
-          animatedSheetState.value !== SHEET_STATE.EXTENDED &&
-          animatedSheetState.value !== SHEET_STATE.FILL_PARENT &&
-          y > 0
-        ) {
-          context.shouldLockInitialPosition = true;
-        } else {
-          context.shouldLockInitialPosition = false;
-        }
-      },
-      onScroll: (_, context) => {
-        /**
-         * if sheet position is extended or fill parent, then we reset
-         * `shouldLockInitialPosition` value to false.
-         */
-        if (
-          animatedSheetState.value === SHEET_STATE.EXTENDED ||
-          animatedSheetState.value === SHEET_STATE.FILL_PARENT
-        ) {
-          context.shouldLockInitialPosition = false;
-        }
-
-        if (animatedScrollableState.value === SCROLLABLE_STATE.LOCKED) {
-          const lockPosition = context.shouldLockInitialPosition
-            ? context.initialContentOffsetY ?? 0
-            : 0;
-          // @ts-ignore
-          scrollTo(scrollableRef, 0, lockPosition, false);
-          scrollableContentOffsetY.value = lockPosition;
-          return;
-        }
-      },
-      onEndDrag: ({ contentOffset: { y } }: NativeScrollEvent, context) => {
-        if (animatedScrollableState.value === SCROLLABLE_STATE.LOCKED) {
-          const lockPosition = context.shouldLockInitialPosition
-            ? context.initialContentOffsetY ?? 0
-            : 0;
-          // @ts-ignore
-          scrollTo(scrollableRef, 0, lockPosition, false);
-          scrollableContentOffsetY.value = lockPosition;
-          return;
-        }
-        if (animatedAnimationState.value !== ANIMATION_STATE.RUNNING) {
-          scrollableContentOffsetY.value = y;
-          rootScrollableContentOffsetY.value = y;
-        }
-      },
-      onMomentumEnd: ({ contentOffset: { y } }: NativeScrollEvent, context) => {
-        if (animatedScrollableState.value === SCROLLABLE_STATE.LOCKED) {
-          const lockPosition = context.shouldLockInitialPosition
-            ? context.initialContentOffsetY ?? 0
-            : 0;
-          // @ts-ignore
-          scrollTo(scrollableRef, 0, lockPosition, false);
-          scrollableContentOffsetY.value = 0;
-          return;
-        }
-        if (animatedAnimationState.value !== ANIMATION_STATE.RUNNING) {
-          scrollableContentOffsetY.value = y;
-          rootScrollableContentOffsetY.value = y;
-        }
-      },
-    });
-
-  const handleSettingScrollable = useHandleSettingScrollable({
-    type,
-    scrollableRef,
-    scrollableContentOffsetY,
-    refreshable,
-  });
-
-  return {
-    scrollableRef,
-    scrollableAnimatedProps,
-    handleScrollEvent,
-    handleSettingScrollable,
-  };
-};
