+++ src/utilities/index.ts
@@ -2,3 +2,4 @@ export { normalizeSnapPoint } from './normalizeSnapPoint';
 export { animate } from './animate';
 export { getKeyboardAnimationConfigs } from './getKeyboardAnimationConfigs';
 export { print } from './logger';
+export { noop, workletNoop } from './noop';
