+++ src/utilities/noop.ts
@@ -1,0 +1,7 @@
+const workletNoop = () => {
+  'worklet';
+};
+
+const noop = () => {};
+
+export { noop, workletNoop };
